Revert "Feature/기사님상세 페이지"
