--- src/services/moverDetailService.ts
@@ -1,36 +1,0 @@
-import { MoverDetail } from '@/services/types/mover';
-import axiosInstance from '@/lib/axiosInstance';
-
-export async function getMoverDetail(
-  moverId: string,
-  cookie?: string | null,
-): Promise<MoverDetail> {
-  const response = await axiosInstance.get(`/movers/${moverId}`, {
-    headers: cookie ? { Cookie: cookie } : undefined,
-  });
-  return response.data.data;
-}
-
-export async function toggleFavorite(
-  moverId: string,
-): Promise<{ isFavorite: boolean; totalCustomerFavorite: number }> {
-  const response = await axiosInstance.post(`/favorites/create/${moverId}`);
-  return {
-    isFavorite: true,
-    totalCustomerFavorite: response.data.totalCustomerFavorite,
-  };
-}
-
-export async function checkFavoriteStatus(
-  moverId: string,
-): Promise<{ isFavorite: boolean; totalCustomerFavorite: number }> {
-  try {
-    const response = await axiosInstance.get(`/favorites/check/${moverId}`);
-    return response.data;
-  } catch (error) {
-    return {
-      isFavorite: false,
-      totalCustomerFavorite: 0,
-    };
-  }
-}
